requirements.txt
# ── API Framework ──────────────────────────────
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.5.0
pydantic-settings==2.1.0

# ── Database ────────────────────────────────────
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# ── Data Processing ────────────────────────────
pandas==1.5.3
numpy==1.24.3

# ── Machine Learning ───────────────────────────
scikit-learn==1.3.2
joblib==1.3.2
xgboost==2.0.3

# ── Testing ────────────────────────────────────
pytest==7.4.3
pytest-cov==4.1.0

# ── Authentication ────────────────────────────
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6

# ── Utilities ───────────────────────────────────
python-dotenv==1.0.0
requests==2.31.0
